requirements.txt
asgiref==3.3.1
boto3==1.16.63
botocore==1.19.63
Django==3.1.6
django-environ==0.4.5
djangorestframework==3.12.2
jmespath==0.10.0
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
